OOO_for_iql/requirements.txt
numpy >= 1.20.2
scipy >= 1.6.0
absl-py >= 0.12.0
gdown >= 3.12.2
tqdm >= 4.60.0
jax[cuda12_pip] == 0.4.18
-f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
jaxlib == 0.4.18
flax == 0.6.10
ml_collections >= 0.1.0
optax >= 0.1.5
tensorboardX == 2.1
tensorflow-probability >= 0.14.1
imageio >= 2.9.0
imageio-ffmpeg >= 0.4.3
d4rl @ git+https://github.com/nakamotoo/D4RL@master#egg=d4rl
cython < 3
wandb >= 0.15.2
tqdm
